Moved delay tasks onto the map square, rather than the location
